% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularPermutationClassifierConstructor.R   
\docType{methods}
\name{loadBCDataset}
\alias{loadBCDataset}
\alias{loadBCDataset,classGeneratorFunction-method}
\alias{loadBCDataset-methods}
\title{MolecularPermutationClassifier high level constructor} 
\usage{
loadBCDataset(Class, libname = c("upp", "nki", "vdx", "mainz", "transbig", 
    "unt"), verbose = getOption("verbose", default = FALSE))  

\S4method{loadBCDataset}{classGeneratorFunction}(Class, libname = c("upp", 
    "nki", "vdx", "mainz", "transbig", "unt"), verbose = getOption("verbose", 
    default = FALSE))  
}
\arguments{
\item{Class}{name of MolecularPermutationClassifier child class to use.}  

\item{libname}{lowercase character with the name of the breastCancerXXX 
database to be loaded. At present, XXX can be "upp", "nki", "vdx", 
"mainz", "transbig" or "unt". See reference for available breast cancer   
citations.}

\item{verbose}{should the user feedback be displayed? By default value is   
"verbose" global option parameter, if present, or FALSE otherwise.}
}
\value{
MolecularPermutationClassifier subclass object with exprs,
annotation and targets slots taken from the libname used.
}
\description{
High level constructor for MolecularPermutationClassifier
subclasses using available Bioconductor's Breast Cancer example datasets. 
}
\examples{
##Using Breast Cancer NKI database, if available, to create a PAM50 class. 
if(requireNamespace("breastCancerNKI")){
    object<-loadBCDataset(Class=PAM50, libname="nki", verbose=TRUE)  
    object

    ##Now we can inspect the object   
    head(exprs(object))      ##The gene expression
    head(annotation(object)) ##The available annotation 
    head(targets(object))    ##The clinical data present in the package  
}

}
\author{
Cristobal Fresno \email{cfresno@bdmg.com.ar}, German A. Gonzalez   
        \email{ggonzalez@bdmg.com.ar}, Andrea S. Llera 
        \email{allera@leloir.org.ar} and Elmer Andres Fernandez    
        \email{efernandez@bdmg.com.ar}
}
\references{
\itemize{
        \item Schroeder M, Haibe-Kains B, Culhane A, Sotiriou C, Bontempi G and 
        Quackenbush J (2011). breastCancerUPP: Gene expression dataset  
        published by Miller et al. [2005] (UPP).. R package version 1.3.1,  
        \url{http://compbio.dfci.harvard.edu/}.
        \item Schroeder M, Haibe-Kains B, Culhane A, Sotiriou C, Bontempi G and 
        Quackenbush J (2011). breastCancerNKI: Genexpression dataset   
        published by van't Veer et al. [2002] and van de Vijver et al.
        [2002] (NKI).. R package version 1.3.1,   
        \url{http://compbio.dfci.harvard.edu/}.
        \item Schroeder M, Haibe-Kains B, Culhane A, Sotiriou C, Bontempi G and 
        Quackenbush J (2011). breastCancerVDX: Gene expression datasets  
        published by Wang et al. [2005] and Minn et al. [2007] (VDX). R
        package version 1.3.1, \url{http://compbio.dfci.harvard.edu/}. 
        \item Schroeder M, Haibe-Kains B, Culhane A, Sotiriou C, Bontempi G and 
        Quackenbush J (2011). breastCancerTRANSBIG: Gene expression dataset  
        published by Desmedt et al. [2007] (TRANSBIG).. R package version   
        1.3.1, \url{http://compbio.dfci.harvard.edu/}.   
        \item Schroeder M, Haibe-Kains B, Culhane A, Sotiriou C, Bontempi G and 
        Quackenbush J (2011). breastCancerMAINZ: Gene expression dataset  
        published by Schmidt et al. [2008] (MAINZ).. R package version   
        1.3.1, \url{http://compbio.dfci.harvard.edu/}.   
        \item Schroeder M, Haibe-Kains B, Culhane A, Sotiriou C, Bontempi G and 
        Quackenbush J (2011). breastCancerUNT: Gene expression dataset  
        published by Sotiriou et al. [2007] (UNT).. R package version   
        1.3.1, \url{http://compbio.dfci.harvard.edu/}.   
}
}
\seealso{
\code{\link{PAM50}} for a complete example.

Other MolecularPermutationClassifier PAM50: \code{\link{PAM50-class}}, 
    \code{\link{filtrate}}
}

