% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularPermutationClassifierGetseters.R   
\docType{methods}
\name{parameters}
\alias{annotation}
\alias{annotation,MolecularPermutationClassifier-method}
\alias{annotation<-}
\alias{annotation<-,MolecularPermutationClassifier,ANY-method}
\alias{classification}
\alias{classification,MolecularPermutationClassifier-method}
\alias{classification-methods}
\alias{exprs}
\alias{exprs,MolecularPermutationClassifier-method}
\alias{exprs<-}
\alias{exprs<-,MolecularPermutationClassifier,ANY-method}
\alias{parameters}
\alias{parameters,MolecularPermutationClassifier-method}
\alias{parameters-methods}
\alias{parameters<-}
\alias{parameters<-,MolecularPermutationClassifier-method}
\alias{parameters<--methods}
\alias{permutation}
\alias{permutation,MolecularPermutationClassifier-method}
\alias{permutation-methods}
\alias{targets}
\alias{targets,MolecularPermutationClassifier-method}
\alias{targets-methods}
\alias{targets<-}
\alias{targets<-,MolecularPermutationClassifier-method}
\alias{targets<--methods}
\title{Accessors for MolecularPermutationClassifier child class slots}   
\usage{
parameters(object)

\S4method{parameters}{MolecularPermutationClassifier}(object)

parameters(object) <- value  

\S4method{parameters}{MolecularPermutationClassifier}(object) <- value  

\S4method{exprs}{MolecularPermutationClassifier}(object)

\S4method{exprs}{MolecularPermutationClassifier,ANY}(object) <- value  

\S4method{annotation}{MolecularPermutationClassifier}(object, ...)   

\S4method{annotation}{MolecularPermutationClassifier,ANY}(object) <- value  

targets(object)

\S4method{targets}{MolecularPermutationClassifier}(object)

targets(object) <- value  

\S4method{targets}{MolecularPermutationClassifier}(object) <- value  

classification(object)

\S4method{classification}{MolecularPermutationClassifier}(object)

permutation(object)

\S4method{permutation}{MolecularPermutationClassifier}(object)
}
\arguments{
\item{object}{MolecularPermutationClassifier subclass object}  

\item{value}{according to the function call:
\itemize{
        \item parameters: named list with at least the following fields:   
            \describe{
                \item{$nPerm}{integer with number of permutations. Default:   
                    1e4L}
                \item{$where}{character with significant value used. Default   
                    value is "fdr".}  
                \item{$pCutoff}{numeric with p-value or fdr cutoff used, i.e., 
                    variable<pCutoff. Default: 0.01}  
                \item{$corCutoff}{numeric with correlation difference between
                    classes cutoff used, i.e., 
                \eqn{|\rho(profile,class_A)-\rho(profile,class_B)|>corCutoff}}
                \item{$keep}{should null distribution simulation values be   
                    kept?. Default: FALSE}  
            }
        \item annotation: data.frame with individual annotations (genes, etc). 
        Minimal compulsory fields are: 
        \describe{
            \item{$probe}{same characters as in row.names(M).}
            \item{$EntrezGene.ID}{integer with NCBI Entrez Data Base.}   
            \item{$NCBI.gene.symbol}{character with gene mnemonic, 
            a.k.a. gene symbol.}  
        }
        \item exprs: matrix with gene exprs profile, where genes are in rows 
        and subjects as columns, a.k.a., \strong{M matrix}.  
        \item targets: data.frame with additional subject data.  
}}

\item{...}{additional parameters according to function call.}   
}
\value{
according to function call one of the following objects:
\item{parameters}{named list see value parameter}
\item{exprs}{matrix with gene exprs profile, where genes are in rows and  
    subjects as columns, a.k.a., \strong{M matrix}.}   
\item{annotation}{data.frame see value parameter} 
\item{classification}{named list with at least the following fields:} 
        \describe{
            \item{$class}{factor with with all possible class levels.}  
        }
\item{permutation}{named list with at least the following fields:} 
        \describe{
            \item{pvalues}{numeric matrix with subjects in row and classes 
            in columns.}   
            \item{$fdr}{numeric matrix with False Discovery Rate correction  
            of pvalues by row.} 
        }
\item{parameters<-}{MolecularPermutationClassifier object with parameters 
    updated slot.}   
\item{exprs<-}{MolecularPermutationClassifier object with exprs updated
    slot.}
\item{annotation<-}{MolecularPermutationClassifier object with annotation 
    updated slot.}   
\item{targets<-}{MolecularPermutationClassifier object with targets 
    updated slot.}   
}
\description{
Slot setters/getters for MolecularPermutationClassifier hierarchy classes   
}
\examples{
##Using pam50centroids package example data
data(pam50centroids)

##Now we can inspect pam50centroids object   
head(exprs(pam50centroids))      ##The gene expression
head(annotation(pam50centroids)) ##The available annotation 
head(targets(pam50centroids))    ##The clinical data present in the package  

##Work with the parameters 
parameters(pam50centroids)       ##Display them
aux<-parameters(pam50centroids)    
aux$keep<-TRUE                   ##Set keep to FALSE  
parameters(pam50centroids)<-aux
parameters(pam50centroids)        

##Also exprs<-, annotation<- and targets<- available functions to update    
##the respective slots  
}
\author{
Cristobal Fresno \email{cfresno@bdmg.com.ar}, German A. Gonzalez   
        \email{ggonzalez@bdmg.com.ar}, Andrea S. Llera 
        \email{allera@leloir.org.ar} and Elmer Andres Fernandez
        \email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{PAM50}} for a complete example.

Other MolecularPermutationClassifier:   
\code{\link{MolecularPermutationClassifier-class}},
    \code{\link{show}}
}

