% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularPermutationClassifierShow.R   
\docType{methods}
\name{show}
\alias{show}
\alias{show,MolecularPermutationClassifier-method}
\title{\code{Show} a MolecularPermutationClassifier subclass object}
\usage{
\S4method{show}{MolecularPermutationClassifier}(object)
}
\arguments{
\item{object}{an object of MolecularPermutationClassifier class hierarchy}   
}
\value{
console messages displaying the class content   
}
\description{
Basic MolecularPermutationClassifier class information display function   
(slots, dimensions, etc).  
}
\examples{
##For an empty object 
object<-PAM50()
object

##Using pam50centroids package example data
data(pam50centroids)
pam50centroids

}
\author{
Cristobal Fresno \email{cfresno@bdmg.com.ar}, German A. Gonzalez   
        \email{ggonzalez@bdmg.com.ar}, Andrea S. Llera 
        \email{allera@leloir.org.ar} and Elmer Andres Fernandez    
        \email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{PAM50}} for a complete example.

Other MolecularPermutationClassifier:   
\code{\link{MolecularPermutationClassifier-class}},
    \code{\link{parameters}}
}

