% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM50Classify.R   
\docType{methods}
\name{classify,PAM50-method}
\alias{classify,PAM50-method}
\title{\code{classify} subjects with PAM50 molecular signature}   
\usage{
\S4method{classify}{PAM50}(object, std = c("none", "scale", "robust",   
    "median")[1], verbose = getOption("verbose", default = FALSE))  
}
\arguments{
\item{object}{a MolecularPermutationClassifier subclass object.} 

\item{std}{character to select standardization alternative "none"   
(default), "scale" and "robust" as in genefu original implementation,
plus the suggested "median" if many subjects are available.}

\item{verbose}{should the user feedback be displayed? By default value is   
"verbose" global option parameter, if present, or FALSE otherwise.}
}
\value{
a PAM50 object with the updated slots:  
\item{@exprs}{updated matrix with the used std parameter.}  
\item{@classification}{
        \describe{
            \item{$subtype}{subject named factor with all classifier possible  
            levels, i.e, "Basal", "Her2", "LumA", "LumB" and "Normal".} 
            \item{$probability}{numeric matrix with subtype class probability   
            for each subject, as in genefu, obtained as the positive   
            proportion of correlation explained by each subtype.}  
            \item{$correlation}{numeric matrix with Spearman's rho
            correlation of each subject to the corresponding PAM50 
            subtypes.}
        }
}
}
\description{
Obtain PAM50 subtype using genefu centroid Spearman's correlation 
implementation. If std=="median" probes with the same mapping are averaged.   
Then, the complete database is center normalized using gene median   
expression. This is done in order to assure selecting the same "gene" to
those in "genefu" library, instead of the most variant probe (default in 
geneid.map), when more than one probe match the same gene. This selection 
is based on probe population variance that could depends on the number of
accounted genes.   
}
\examples{
##Using pam50centroids package example data
data(pam50centroids)

##Get the original PAM50 calls using genefu implementation 
pam50centroids<-classify(pam50centroids, std="none", verbose=TRUE)  
classification(pam50centroids)
}
\author{
Cristobal Fresno \email{cfresno@bdmg.com.ar}, German A. Gonzalez   
        \email{ggonzalez@bdmg.com.ar}, Andrea S. Llera 
        \email{allera@leloir.org.ar} and Elmer Andres Fernandez    
        \email{efernandez@bdmg.com.ar}
}
\references{
\enumerate{
        \item Haibe-Kains B, Schroeder M, Bontempi G, Sotiriou C and   
        Quackenbush J, 2014, genefu: Relevant Functions for Gene 
        Expression Analysis, Especially in Breast Cancer. R package 
        version 1.16.0, \url{www.pmgenomics.ca/bhklab/}      
        \item Perou CM, Sorlie T, Eisen MB, et al., 2000, Molecular portraits 
        of human breast tumors. Nature 406:747-752.   
        \item Perou CM, Parker JS, Prat A, Ellis MJ, Bernard PB., 2010, 
        Clinical implementation of the intrinsic subtypes of  
        breast cancer, The Lancet Oncology 11(8):718-719.   
}
}
\seealso{
\code{\link{PAM50}} for a complete example.

Other PAM50: \code{\link{as}},  
    \code{\link{filtrate,PAM50-method}},
    \code{\link{pam50centroids}},
    \code{\link{permutate,PAM50-method}},
    \code{\link{subjectReport,PAM50-method}},
    \code{\link{subtypes,PAM50-method}}
}

