% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM50Filtrate.R   
\docType{methods}
\name{filtrate,PAM50-method}
\alias{filtrate,PAM50-method}
\title{\code{filtrate} centroid genes from PAM50 classification}   
\usage{
\S4method{filtrate}{PAM50}(object, verbose = getOption("verbose", default =   
    FALSE))
}
\arguments{
\item{object}{a PAM50 object.}  

\item{verbose}{should the user feedback be displayed? By default value is   
"verbose" global option parameter, if present, or FALSE otherwise.}
}
\value{
MolecularPermutationClassifier subclass with updated slots:
\item{@exprs}{only rows required by the classifier.}   
\item{@annotation}{consistent with exprs rows and only "probe",  
        "EntrezGene.ID" and "NCBI.gene.symbol" annotation fields.}
}
\description{
Remove exprs rows not required by MolecularPermutationClassifier subclasses 
to classify samples, in this case PAM50. This means to only keep genes with   
valid EntrezGeneID, i. e., not NA and present in PAM50 signature centroids. 
In addition, annotation slot will only keep "probe", "EntrezGene.ID" and   
"NCBI.gene.symbol" fields required by genefu's intrinsic.cluster.predict   
function.
}
\examples{
##Using pam50centroids package example data
data(pam50centroids)
pam50centroids
pam50centroids<-filtrate(pam50centroids, verbose=TRUE)   
pam50centroids

##Using Breast Cancer NKI database, if available.  
if(requireNamespace("breastCancerNKI")){
    object<-loadBCDataset(Class=PAM50, libname="nki", verbose=TRUE)  
    object
    object<-filtrate(object, verbose=TRUE)   
    object
}

}
\author{
Cristobal Fresno \email{cfresno@bdmg.com.ar}, German A. Gonzalez   
        \email{ggonzalez@bdmg.com.ar}, Andrea S. Llera 
        \email{allera@leloir.org.ar} and Elmer Andres Fernandez
        \email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{PAM50}} for a complete example.

Other PAM50: \code{\link{as}},  
    \code{\link{classify,PAM50-method}},
    \code{\link{pam50centroids}},
    \code{\link{permutate,PAM50-method}},
    \code{\link{subjectReport,PAM50-method}},
    \code{\link{subtypes,PAM50-method}}
}

