% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM50Report.R   
\docType{methods}
\name{subjectReport,PAM50-method}
\alias{databaseReport,PAM50-method}
\alias{subjectReport,PAM50-method}
\alias{summary,PAM50-method}
\title{PAM50 permutation test results reports}
\usage{
\S4method{subjectReport}{PAM50}(object, subject)   

\S4method{databaseReport}{PAM50}(object, fileName, ...,  
    verbose = getOption("verbose", default = TRUE))   

\S4method{summary}{PAM50}(object, ...)   
}
\arguments{
\item{object}{a PAM50 object.}  

\item{subject}{integer to select the appropriate subject to report.} 

\item{fileName}{character with the name of the pdf report file to save.}  

\item{...}{additional parameters for pdf function call.}   

\item{verbose}{should the user feedback be displayed? By default value is   
"verbose" global option parameter, if present, or FALSE otherwise.}
}
\value{
depending on function call: 
\item{subjectReport}{a grid.arrange object.}  
\item{databaseReport}{a pdf file with database summary and  
        \code{subjectReport}s.}
\item{summary}{Console summary statistics plus a data.frame}   
}
\description{
\code{subjectReport} is basically a grid.arrange object which basically 
consists of three main parts: a summary table, a two row ggplot2 
facet_wrap with scatter ggplots (Wickham 2009) of subject expression and   
PAM50 centroids (Perou et al. 2000 & 2010) and a textGrob with the
simulation parameter used. Particularly: 
\describe{
        \item{tableGrob}{ with the following fields:
            \describe{
                \item{$Summary}{subject name, PAM50 and Permuted subtype}   
                \item{$Fields}{for the five PAM50 subtypes:    
                    \itemize{
                        \item Correlation: PAM50 centroid correlation with   
                            observed subject exprs.  
                        \item p-value: permutation p-value obtained using   
                            the simulation.   
                        \item FDR: adjusted p-value using False   
                        Discovery Rate.   
                    }
                }
            }
        }
        \item{ggplot facet_wrap}{two rows to display scatter subject exprs vs
        PAM50 centroids, in addition to a the linear regression fix. If  
        subject, has an unique subtype, then the graph is in red. In 
        addition, if simulated permutations were run with keep=TRUE 
        option, then null distribution boxplots are plotted with 
        observed correlations as a big round point.}  
        \item{textGrob}{the permutation @parameter slot used in the  
        simulation.}    
}
}
\details{
\code{summary} it basically prints descriptive data of PAM50 dataset,
the test parameters used, a frequency table of PAM50 Subtypes and a 
contingency table with Classes vs PAM50 Subtypes.  

\code{databaseReport} basically is a pdf report where the first page is a 
global summary of the database, i.e., a \code{summary} contingency table   
of permutation test classes against original PAM50 subtypes results. The   
following pages are the database respective \code{subjectReport} outputs. 
}
\examples{
##Using pam50centroids package example data
data(pam50centroids)
pam50centroids

##This object has already run filtrate, classify and permutate. So, now  
##we can obtain some reports:
##1) database summary  
summary(pam50centroids)

##2)Individual subject report. If keep=FALSE boxplot panel is not available   
subjectReport(pam50centroids, subject=1)##Basal subtype  
subjectReport(pam50centroids, subject=1)##Her2 subtype  

##3) complete database report 
#databaseReport(pam50centroids, fileName="PAM50.pdf", verbose=TRUE)  
}
\author{
Cristobal Fresno \email{cfresno@bdmg.com.ar}, German A. Gonzalez   
        \email{ggonzalez@bdmg.com.ar}, Andrea S. Llera 
        \email{allera@leloir.org.ar} and Elmer Andres Fernandez
        \email{efernandez@bdmg.com.ar}
}
\references{
\enumerate{
        \item Perou CM, Sorlie T, Eisen MB, et al., 2000, Molecular portraits 
        of human breast tumors. Nature 406:747-752.   
        \item Perou CM, Parker JS, Prat A, Ellis MJ, Bernard PB., 2010, 
        Clinical implementation of the intrinsic subtypes of breast 
        cancer, The Lancet Oncology 11(8):718-719.
        \item Wickham H, ggplot2: elegant graphics for data analysis.
            Springer New York, 2009. 
}
}
\seealso{
\code{\link{PAM50}} for a complete example.

Other PAM50: \code{\link{as}},  
    \code{\link{classify,PAM50-method}},
    \code{\link{filtrate,PAM50-method}},
    \code{\link{pam50centroids}},
    \code{\link{permutate,PAM50-method}},
    \code{\link{subtypes,PAM50-method}}

Other PAM50: \code{\link{as}},  
    \code{\link{classify,PAM50-method}},
    \code{\link{filtrate,PAM50-method}},
    \code{\link{pam50centroids}},
    \code{\link{permutate,PAM50-method}},
    \code{\link{subtypes,PAM50-method}}

Other PAM50: \code{\link{as}},  
    \code{\link{classify,PAM50-method}},
    \code{\link{filtrate,PAM50-method}},
    \code{\link{pam50centroids}},
    \code{\link{permutate,PAM50-method}},
    \code{\link{subtypes,PAM50-method}}
}

